If the ‘Segment for LGD’ field has the value ‘Secured’, then the ‘Collateral Type’ should be ‘Fixed Deposit (Cash Collateral)’. If the ‘Segment for LGD’ field has the value ‘Unsecured’, then the ‘Collateral Type’ should be ‘Personal Guarantee’.
If the ‘Product Category’ has the value ‘Credit Cards’, then the ‘Collateral Type’ can be either empty or ‘Fixed Deposit (Cash Collateral)’. If the ‘Product Category’ has the value ‘Gold Loan’, then the ‘Collateral Type’ should be ‘Gold’. If the ‘Product Category’ has the value ‘Housing Loan’, then the ‘Collateral Type’ should be ‘Property’.

For the combination Product Category = ‘Lease’, Segment for PD = ‘LE-Motor Car’, and Segment for LGD = ‘Non-Hybrid’, the Collateral Type should be ‘Car Non-Hybrid’.

For the combination Product Category = ‘Lease’, Segment for PD = ‘LE-Other’, and Segment for LGD = ‘LE-Other’, the Collateral Type should be ‘Machinery’.

For the combination Product Category = ‘Lease’, Segment for PD = ‘LE-Motor Car’, and Segment for LGD = ‘Hybrid’, the Collateral Type should be ‘Car Hybrid’.

If the ‘Product Category’ has the value ‘Overdraft’ or ‘Short Term Loan’, then the ‘Collateral Type’ should be empty.

'‘Collateral Type’ should be consistent across all periods for a given facility.